reajustando ejercicio de blogs
