+++ student-blogs/fjquant-2458/index.html
@@ -5,10 +5,14 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="styles/main.scss">
+    <script defer type="module" src="./Fechas.mjs"></script>
     <title>Blog</title>
 </head>
 
 <body>
+    <header>
+        <h1>Playas de Nicaragua</h1>
+    </header>
     <div class="container">
         <nav>
             <UL>
@@ -22,23 +26,54 @@
         <div class="principal">
             <img src="/Imagenesblog/Corn-Island.jpg" alt="isla corn island">
             <h3>Corn Island</h3>
-            <p><strong>Island</strong> es un archipiélago ubicado en la costa sur de este país, conformado por dos islas del mismo nombre con diferencias de que una es más grande que la otra, y están separadas a 16 Km de distancia. Para llegar a ella se puede hacer por avión a través el Aeropuerto Internacional de Managua lo cual son 45min de viaje, o por ferri desde Bluefields recorrido que se toma un tiempo aproximado de 02 horas.</p>
+            <time datetime="2023-04-05"></time>
+            <p><strong>Corn Island</strong> es un archipiélago ubicado en la costa sur de este país, conformado por dos
+                islas del mismo nombre con diferencias de que una es más grande que la otra, y están separadas a 16 Km
+                de distancia. Para llegar a ella se puede hacer por avión a través el Aeropuerto Internacional de
+                Managua lo cual son 45min de viaje, o por ferri desde Bluefields recorrido que se toma un tiempo
+                aproximado de 02 horas.</p>
+            <a href="#">más...</a>
         </div>
         <div class="articulo">
             <img src="/Imagenesblog/Playa-Colorado.jpg" alt="Playa Colorado">
             <h3>Playa Colorado</h3>
-            <P>Unicada en tola, pueblo al sur de la ciudad de Rivas, reconocida por ser sede del Campeonato Mundial Master en el año 2012, considerada una de las mejores playas de Nicaragua, visitada por más de 17 mil turistas en temporadas como semana santa y carnavales, Playa Colorado es popular entre los surfistas por tener olas hasta de 2 metros de alto.</P>
+            <time datetime="2023-02-20"></time>
+            <P>Unicada en tola, pueblo al sur de la ciudad de Rivas, reconocida por ser sede del Campeonato Mundial
+                Master en el año 2012, considerada una de las mejores playas de Nicaragua, visitada por más de 17 mil
+                turistas en temporadas como semana santa y carnavales, Playa Colorado es popular entre los surfistas por
+                tener olas hasta de 2 metros de alto.</P>
+            <a href="#">más...</a>
         </div>
         <div class="articulo">
             <img src="/Imagenesblog/Playa-El-Coco.jpg" alt="Playa El Coco">
             <h3>Playa El Coco</h3>
-            <P>A tan solo 159 km de la capital, al sur de la ciudad Puerto de San Juan, nos situamos en la hermosa Playa El Coco, donde su ubicación ideal, aguas mansas y exuberante vegetación, hace de ellas un destino ideal para vacacionar.</P>
+            <time datetime="2023-01-21"></time>
+            <p>A tan solo 159 km de la capital, al sur de la ciudad Puerto de San Juan, nos situamos en la hermosa Playa
+                El Coco, donde su ubicación ideal, aguas mansas y exuberante vegetación, hace de ellas un destino ideal
+                para vacacionar.</P>
+            <a href="#">más...</a>
         </div>
         <div class="articulo">
             <img src="/Imagenesblog/San-Juan-del-Sur.jpg" alt="Playa en San Juan del Sur">
             <h3>San Juan del Sur</h3>
-            <P>Ubicada en un pequeño pueblo con el mismo nombre, San Juan del Sur es una bahía en forma de herradura, rodeada de hermosa vegetación, causa deleite por su encanto propio, clima agradable y ambiente local, esta bahía ofrece increíbles destinos para la práctica de deportes extremos como el rappel, surf, kayak, entre otros.</P>
-        </div>
+            <time datetime="2022-12-15"></time>
+            <P>Ubicada en un pequeño pueblo con el mismo nombre, San Juan del Sur es una bahía en forma de herradura,
+                rodeada de hermosa vegetación, causa deleite por su encanto propio, clima agradable y ambiente local,
+                esta bahía ofrece increíbles destinos para la práctica de deportes extremos como el rappel, surf, kayak,
+                entre otros.</P>
+                <a href="#">más...</a>
+            </div>
+        <div class="articulo">
+            <img src="/Imagenesblog/playa-poneloya.jpg" alt="Playa Poneloya">
+            <h3>Playa Poneloya</h3>
+            <time datetime="2022-11-15"></time>
+            <P>Poneloya es un pueblo pesquero de una sola calle paralela al mar; y un centro de casas de veraneo. La
+                playa es irregular y de arena oscura, y deja ver palmeras que completan su aspecto de playa tropical.
+
+                Es conocida por los nicaragüenses como una playa en cuyas aguas se presentan fuertes corrientes y olas
+                que rompen con fuerza sobre la arena.</P>
+                <a href="#">más...</a>
+            </div>
     </div>
 </body>
