+++ student-blogs/fjquant-2458/styles/desktop.scss
@@ -4,22 +4,18 @@
         display: grid;
         grid-template-columns: repeat(3 1fr);
         grid-template-rows: repeat(4 1fr);
-        gap: 20px;
+        gap: 20px;;
     }
 
 
     img {
         width: 100%;
-        height: auto;
+        height: 200px;
     }
 
     .principal {
         grid-column: 1 / 3;
-        grid-row: 2/ 4;
+        grid-row: 2/ 3;
         align-items: center;
-    }
-
-    div:last-child{
-        grid-column: 1 / 4;
     }
 }
