+++ student-blogs/fjquant-2458/styles/main.scss
@@ -1,7 +1,3 @@
 @use "base.scss";
 @use "desktop.scss";
 @use "variables.scss";
-
-body{
-    background-color: antiquewhite;
-}
