+++ student-blogs/fjquant-2458/styles/variables.scss
@@ -1,4 +1,7 @@
+:root{
+--color-Champagne: #F7E7CE;
+--color-Champagne-Ligth: #eed9b6;
 
-$color-main: #faebd7;
+--color-Danube: #2377a2;
 
-$font-color: #6699CC
+}
